Rename idle no-op tool
