+++ profiles/chess_coach/tools.txt
@@ -3,6 +3,6 @@ stop_dance
 play_emotion
 stop_emotion
 camera
-do_nothing
+idle_do_nothing
 head_tracking
 move_head
