+++ profiles/example/tools.txt
@@ -5,7 +5,7 @@ stop_dance
 play_emotion
 stop_emotion
 # camera
-# do_nothing
+# idle_do_nothing
 # head_tracking
 # move_head
