+++ profiles/tedai/tools.txt
@@ -3,5 +3,5 @@ stop_dance
 play_emotion
 stop_emotion
 camera
-do_nothing
+idle_do_nothing
 move_head
